Implement product repository.
The main reason behind this repository is that when adding new product with non-existing brand or provider or category exception should be thrown. If this thing was to be done with dbcontext there would be a lot of repetitive code for checking that product is added correctly. Also, by default repository gets a product with information about its brand, images and category.